requirements.txt
Flask==1.1.1
requests==2.23.0
pymysql
werkzeug.security
flaskext.mysql
flask-mysqldb
flask_table
